Add Makefile to run tests; remove commented out code
Added a simple Makefile that uses lunit to run the test suite.
Typing 'make' or 'make test' will run all the tests in the tests/
directory.

Removed commented out print() statements from code.

One test case will output an error on the console which is confusing
to users. Changed the message so that users understand what is
happening.