ngx sender: try to process partial responses as normal
Depending on the server/proxy configuration, it happens that a TCP reset
is sent right after a successful response. When that happens, the `:read()`
method returns an error with a "partial" response (which is full in this
case).

Now we try to process that partial buffer as normal and see if it is well
formed HTTP.

Fixes #30